ugly hack :(
